LaTeX: use again L(eft), not J(ustified), as default for table columns
The default was changed from L to J at #3435. But, using J means that
comma separated items (as in function signatures) will have white space
between words extended, giving non aesthetic results. The L is
appropriate for table cells with a paragraph or paragraphs, but I
think this is not majority of use cases.