LaTeX: add support for named destinations (without prefix) in PDFs
Allow to link externally to some permanent PDF ids, like this

    http://example.com/doc.pdf#nameddest=my-target

where my-target is chosen label in source (i.e. the document name is not
added as prefix). This allows external linking in a way stable against
renaming of sections or moving documentation from one document to the
other in the same project.

(The "nameddest=" part is even optional in some viewer)

Closes: #6276
Closes: #8807