Experimental \sphinxbreakablebox via sphinxbreakablebox.sty
See discussion at #11224

This is about a variant of Sphinx 6.2.0 \sphinxbox, which will break
across lines.  The way this is done is to split the input into
characters, handling encountered macros in various more or less apt
ways in passing and box them separately adapting the shape as first
and last must be handled especially.

Works with one-character input, even empty input.  See the file for
more explanations.

Decision has been made not to merge into Sphinx.

This commit puts the experimental code in a separate package in order
to make using it as easy as possible.  Simply grab the file, put
it at some place where TeX can find it or in your project with
latex_additional_files, and add \usepackage{sphinxbreakablebox} to
preamble.

Use at own risk.  It is not excluded that renamings of Sphinx
internals could at some future point break the file, but I will try to
sync with upstream as long as I contribute maintenance to Sphinx
LaTeX.