(DOES NOT WORK) Let sphinxVerbatim work with TABs
This makes our best to let fancyvrb "obeytabs" feature work for Sphinx,
and will work if TABs are only at "top level", however there is core
deficiency of the "obeytabs" feature of fancyvrb that TABs in Verbatim
can not be inside a macro argument, as is mandatory for Sphinx with
commandchars=\\\{\} to render the Pygments highlighting mark-up macros.

So I am keeping this branch only for not wasting the time which was
devoted to exploratory work, but there is no way we can do this without
redefining ourselves entirely the way \FV@Tab (\FV@TrueTab) acts, if we
want to keep a LaTeX solution.

A much much easier way would be to actually have all TABs converted to
spaces, as is what happens with code-block contents, also in the case
of literalinclude so that we do not have to solve ourself a grave
upstream deficiency of fancyvrb.sty.